Move mocks to utils folder and fix failing test
